+++ tests/test_loader.lua
@@ -50,7 +50,7 @@ end
 assert(run.loader.mods.minimal_cheats, "SilverShadow was not discovered")
 assert(#run.errors == 0, "loader errors: " .. table.concat(run.errors, "; "))
 assert(run.loader.exports.minimal_cheats, "SilverShadow exports were not published")
-assert(run.loader.exports.minimal_cheats.silvershadow.version == "2.0.0",
+assert(run.loader.exports.minimal_cheats.silvershadow.version == "2.0.1",
   "wrong runtime export version")
 assert(run.data.constants.bagSize == 999, "distinct bag-type capacity was not merged")
 assert(run.data.field.pcItemCap == 999, "PC item-stack capacity was not merged")